Moved project types to project-types.ts
In preparation for implementing connectCommand()
